Fixing some weird randomness issue of java.util.Random

+++ gemp-lotr/gemp-lotr-server/src/main/java/com/gempukku/lotro/draft2/builder/DraftChoiceBuilder.java
@@ -99,7 +99,7 @@ public class DraftChoiceBuilder {
             private List<CardCollection.Item> getCards(long seed, int stage) {
                 Random rnd = getRandom(seed, stage);
                 // Fixing some weird issue with Random
-                float thisFixesRandomnessForSomeReason = rnd.nextFloat();
+                float thisFixesRandomnessForSomeReason = rnd.nextInt();
                 final List<CardCollection.Item> cards = new ArrayList<CardCollection.Item>(possibleCards);
                 Collections.shuffle(cards, rnd);
                 return cards;
